Fixed bug where back-tick system calls where failing.
When making a system call using back-ticks, eg: /Users/brendanboesen/Documents/projects/ci_reporter, the call was failing because  was set to nil.
This turned out to be a bug in the initialization of the Delegate where the delegate instance (currently nil) was passed to the 'assign' block instead of the captured IO stream.