Feature: RSpec 3 support
RSpec 3 has completely overhauled the formatter system using a
notification system that allows developers to optionally register
interest in particular notifications instead of implementing them all.

Please read: http://myronmars.to/n/dev-blog/2013/07/the-plan-for-rspec-3
for more information.

Currently the RSpec class is toggled depending on the version of RSpec
however there are currently no tests in place for the RSpec 3 formatter
class.